Fix: Add the IT case back
